requirements.txt
chess
pyserial